requirements.txt
QtPy>=2.4.1
PyQt6>=6.6.0
gdal>=3.8.0
rasterio>=1.3.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
pyproj>=3.6.0
requests>=2.31.0
matplotlib>=3.7.0
scikit-image>=0.21.0
elevation>=1.1.3
python-dotenv>=1.0.0
